- Implemented feature request 7841: Added possibility to provide a TTL of 0 to   make cache items last endlessly.
git-svn-id: https://svn.apache.org/repos/asf/incubator/zetacomponents/trunk@970036 13f79535-47bb-0310-9956-ffa450edef68

+++ Cache/src/storage/file.php
@@ -174,7 +174,7 @@ abstract class ezcCacheStorageFile extends ezcCacheStorage
             return false;
         }
         // Cached data outdated, purge it.
-        if ( $this->calcLifetime( $filename ) > $this->options['ttl'] ) 
+        if ( $this->calcLifetime( $filename ) > $this->options['ttl'] && $this->options['ttl'] !== 0 ) 
         {
             $this->delete( $id, $attributes );
             return false;
